Implement Gmail thread analysis storage and enhance ConversationThreads component
- Added a new API endpoint to store Gmail thread analysis, ensuring proper error handling and validation for user input.
- Enhanced the ConversationThreads component with functionality to trigger thread analysis, display loading states, and extract insights for better user experience.
- Refactored PartnershipDetailPanel to streamline data handling and integrate with the updated ConversationThreads component, removing redundant analysis logic.

These changes improve the overall functionality and user experience related to Gmail thread analysis and partnership management.
